# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Chat API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface GeminiConversation {
  id: number;
  title: string;
  createdAt: string;
  updatedAt: string;
  messageCount: number;
}

export interface GeminiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface GeminiConversationInput {
  title?: string;
}

export interface GeminiMessageInput {
  /** @minLength 1 */
  content: string;
}

export interface GeminiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  updatedAt: string;
  messages: GeminiMessage[];
}

export interface GeminiTitleUpdate {
  /** @minLength 1 */
  title: string;
}

export interface GeminiError {
  error: string;
}