# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Chat API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: gemini
    description: Gemini AI chat and image operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /gemini/conversations:
    get:
      operationId: listGeminiConversations
      tags: [gemini]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeminiConversation"
    post:
      operationId: createGeminiConversation
      tags: [gemini]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeminiConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiConversation"

  /gemini/conversations/{id}:
    get:
      operationId: getGeminiConversation
      tags: [gemini]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiError"
    delete:
      operationId: deleteGeminiConversation
      tags: [gemini]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiError"

  /gemini/conversations/{id}/messages:
    get:
      operationId: listGeminiMessages
      tags: [gemini]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeminiMessage"
    post:
      operationId: sendGeminiMessage
      tags: [gemini]
      summary: Send a message and receive an AI response (SSE stream)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeminiMessageInput"
      responses:
        "200":
          description: SSE stream of assistant response chunks
          content:
            text/event-stream: {}

  /gemini/conversations/{id}/title:
    patch:
      operationId: updateGeminiConversationTitle
      tags: [gemini]
      summary: Update conversation title
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeminiTitleUpdate"
      responses:
        "200":
          description: Updated conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiConversation"
        "404":
          description: Not found

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    GeminiConversation:
      type: object
      required: [id, title, createdAt, updatedAt, messageCount]
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        messageCount:
          type: integer

    GeminiMessage:
      type: object
      required: [id, conversationId, role, content, createdAt]
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time

    GeminiConversationInput:
      type: object
      properties:
        title:
          type: string

    GeminiMessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1

    GeminiConversationWithMessages:
      type: object
      required: [id, title, createdAt, updatedAt, messages]
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/GeminiMessage"

    GeminiTitleUpdate:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1

    GeminiError:
      type: object
      required: [error]
      properties:
        error:
          type: string