# lib/api-zod/src/generated/types/geminiConversationWithMessages.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Chat API
 * OpenAPI spec version: 0.1.0
 */
import type { GeminiMessage } from './geminiMessage';

export interface GeminiConversationWithMessages {
  id: number;
  title: string;
  createdAt: Date;
  updatedAt: Date;
  messages: GeminiMessage[];
}