# lib/api-zod/src/generated/types/geminiMessageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Chat API
 * OpenAPI spec version: 0.1.0
 */

export interface GeminiMessageInput {
  /** @minLength 1 */
  content: string;
}